Add a basic disk drive implementation
Disk drives can currently only support floppy drives (adding support for
computers could come in the future).

The id of the floppy is provided as a config option (id) which is set
when creating the peripheral. Files are read and written to
<data_dir>/computer/disk/<id> (like ComputerCraft). Floppy labela are
not currently persisted, but are shared across all computers: disks
with the same ID will have the same label across the emulator.