Remove async peripheral setters
Historically CC used to call IPeripheral.{attach,detach} on the computer
thread. This meant that peripherals would be in a weird state of being
accessible with peripheral.call, but not actually attached.

In 4df252f7564873c03f61ca3390a64140f798dc92, we added a workaround for
this, which waited until the attach had actually happened before
returning from the ccemux functions.

However, this is no longer needed. As of [this CC:T commit][1],
IPeripheral.{attach,detach} are called immediately.

[1]: https://github.com/cc-tweaked/CC-Tweaked/commit/474f571798722e5964a0af2eb8805ab117295835