Fix several issues with asset dir option
In 4d82b486e905fa8e41ba6b0fa285050ac3595998 we changed the asset
directory to work from the root, rather than the ./assets (or ./data)
directory. However, we failed to update the actual CLI (or
documentation), which meant we were searching for files in the root of
the ccemux directory, rather than $DATA_DIR/data.

This commit:
 - Updates the CLI switch to default to $DATA_DIR/data.
 - Create $ASSET_DIR/computercraft/lua, rather than .../assets/...